backend/requirements.txt
# ── Core ──
Django>=5.0
djangorestframework
django-filter
django-cors-headers
python-dotenv

# ── API Docs ──
drf-spectacular
drf-spectacular-sidecar

# ── Dev tools ──
black
isort
flake8

# ── Tests ──
pytest
pytest-django
